[ads] Cleanup size_t usage

+++ components/brave_ads/core/internal/account/tokens/confirmation_tokens/confirmation_tokens_util.cc
@@ -75,7 +75,7 @@ bool ConfirmationTokensIsEmpty() {
       .IsEmpty();
 }
 
-int ConfirmationTokenCount() {
+size_t ConfirmationTokenCount() {
   return ConfirmationStateManager::GetInstance()
       .GetConfirmationTokens()
       .Count();
